requirements.txt
coverage==4.5.3
dj-database-url==0.5.0
Django==2.2.2
django-allauth==0.39.1
django-db-queue==1.2.0
django-environ==0.4.5
django-log-request-id==1.3.2
django-rest-auth==0.9.5
django-simple-robots==1.0.0
djangorestframework==3.9.2
flake8==3.7.7
mypy==0.701
psycopg2==2.8.2
waitress==1.3.0
whitenoise==4.1.2
raven==6.10.0
unittest-xml-reporting==2.5.1

django-thumborstorage==1.11.1
libthumbor==1.3.2

django-enforce-host==1.0.1
